modules/server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "karrio_server_modules"
version = "2026.1.29"
description = "Multi-carrier shipping API consolidated server modules (core + graph + admin)"
readme = "README.md"
requires-python = ">=3.11"
license = "LGPL-3.0"
authors = [
    {name = "karrio", email = "hello@karrio.io"}
]
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "karrio",
    "psycopg2-binary",
    "django-health-check",
    "dnspython",
    "psutil",
    "pyyaml",
    "Jinja2",
    "django-filter",
    "strawberry-graphql",
]

[project.urls]
Homepage = "https://github.com/karrioapi/karrio"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.packages.find]
exclude = ["tests.*", "tests"]
namespaces = true

[tool.setuptools.package-data]
"karrio.server.user" = ["templates/**/*"]
"karrio.server.graph" = ["templates/**/*"]
"*" = ["*.html", "*.txt", "*.css", "*.js", "*.png", "*.jpg", "*.gif", "*.ico"]
